+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 0.4.9
 
 - fixed bug that caused symbols to persist even after flow was closed
+- also, forgot to remove my console.logs -.-
 
 ### 0.4.8