+++ RELEASE_NOTES.md
@@ -1,3 +1,4 @@
 ### 0.4.9
 
 - fixed bug that caused symbols to persist even after flow was closed
+- also, forgot to remove my console.logs -.-